+++ src/scad_tarmi/core.clj
@@ -5,6 +5,7 @@
 ;; CONSTANTS AND SHORTCUTS ;;
 ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
 
+(def abs #(max % (- %)))
 (def √ #(Math/sqrt %))
 (def sin #(Math/sin %))
 (def cos #(Math/cos %))
@@ -16,6 +17,13 @@
 ;;;;;;;;;;;;;;;;;;;;;;;;;
 ;; INTERFACE FUNCTIONS ;;
 ;;;;;;;;;;;;;;;;;;;;;;;;;
+
+;; The following functions convert between two ways to measure an angle:
+;; - Measured counterclockwise from the x axis, as is typical in trigonometry
+;;   and represents the behaviour of OpenSCAD’s rotate function.
+;; - Measured clockwise from the y axis.
+(defn counterclockwise-from-x-axis [θ] (mod (+ (- τ θ) (/ π 2)) τ))
+(defn clockwise-from-y-axis        [θ] (mod (- (- τ θ) (/ π 2)) τ))
 
 (defn long-hex-diagonal
   "The “long diagonal” of a hexagon, computed from the “short” or flat-to-flat